Refactor macros for smaller file length
